douban_oss/requirements.txt
Scrapy==1.5.0
oss2==2.4.0
Pillow==5.1.0
